--- src/main/java/ru/yandex/practicum/filmorate/dao/FriendDao.java
@@ -1,4 +1,0 @@
-package ru.yandex.practicum.filmorate.dao;
-
-public interface FriendDao {
-}
