+++ src/main/java/ru/yandex/practicum/filmorate/storage/jdbc/AbstractDao.java
@@ -1,0 +1,13 @@
+package ru.yandex.practicum.filmorate.storage.jdbc;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.JdbcTemplate;
+
+public abstract class AbstractDao {
+    protected final JdbcTemplate jdbcTemplate;
+
+@Autowired
+    protected AbstractDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+}
