+++ src/main/java/ru/yandex/practicum/filmorate/storage/jdbc/UserDao.java
@@ -1,0 +1,4 @@
+package ru.yandex.practicum.filmorate.storage.jdbc;
+
+public interface UserDao {
+}
